Shared lib support
