+++ plugins/DB.cpp
@@ -1,24 +1,7 @@
-#include <libstuff/libstuff.h>
-#include "../BedrockPlugin.h"
+#include "DB.h"
 
 #undef SLOGPREFIX
 #define SLOGPREFIX "{" << node->name << ":" << getName() << "} "
-
-// Declare the class we're going to implement below
-class BedrockPlugin_DB : public BedrockPlugin {
-  public:
-    virtual string getName() { return "DB"; }
-    virtual void initialize(const SData& args) { _args = args; }
-    virtual bool peekCommand(BedrockNode* node, SQLite& db, BedrockNode::Command* command);
-    virtual bool processCommand(BedrockNode* node, SQLite& db, BedrockNode::Command* command);
-
-  private:
-    // Attributes
-    SData _args;
-};
-
-// Register for auto-discovery at boot
-BREGISTER_PLUGIN(BedrockPlugin_DB);
 
 bool BedrockPlugin_DB::peekCommand(BedrockNode* node, SQLite& db, BedrockNode::Command* command) {
     // Pull out some helpful variables
