+++ plugins/DB.h
@@ -1,0 +1,15 @@
+#include <libstuff/libstuff.h>
+#include "../BedrockPlugin.h"
+
+// Declare the class we're going to implement below
+class BedrockPlugin_DB : public BedrockPlugin {
+  public:
+    virtual string getName() { return "DB"; }
+    virtual void initialize(const SData& args) { _args = args; }
+    virtual bool peekCommand(BedrockNode* node, SQLite& db, BedrockNode::Command* command);
+    virtual bool processCommand(BedrockNode* node, SQLite& db, BedrockNode::Command* command);
+
+  private:
+    // Attributes
+    SData _args;
+};
